pyproject.toml
[project]
name = "credit-risk"
version = "0.1.0"
description = "AI Credit Risk Decision System — PD modeling, decision engine, explainability, and monitoring"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # --- Data & Feature Engineering ---
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pyarrow>=15.0.0",           # Parquet / Arrow columnar storage
    "polars>=0.20.0",             # Fast DataFrame alternative for large feature jobs
    "scikit-learn>=1.4.0",        # Preprocessing, calibration, metrics
    "feature-engine>=1.6.0",      # WoE encoding, outlier imputation, missing indicators

    # --- ML Models ---
    "lightgbm>=4.3.0",            # Primary production model
    "xgboost>=2.0.0",             # Challenger / ensemble component
    "catboost>=1.2.0",            # Categorical feature handling alternative
    "optuna>=3.6.0",              # Hyperparameter optimization (Bayesian)

    # --- Calibration & Explainability ---
    "shap>=0.46.0",               # SHAP TreeExplainer — 0.46+ uses numba>=0.56 (Py3.11 safe)
    "numba>=0.56.0",              # Force modern numba with Py3.11 wheels (excludes old 0.53)
    "lime>=0.2.0",                # LIME for cross-validation of local explanations

    # --- Fairness ---
    "fairlearn>=0.10.0",          # Fairness metrics + equalized odds post-processing

    # --- API / Serving ---
    "fastapi>=0.111.0",           # Inference API
    "uvicorn[standard]>=0.29.0",  # ASGI server
    "pydantic>=2.7.0",            # Request/response validation
    "redis>=5.0.0",               # Online feature store cache client
    "httpx>=0.27.0",              # Async HTTP client for bureau/bank API calls

    # --- Experiment Tracking & Model Registry ---
    "mlflow>=2.13.0",             # Experiment tracking, model registry, serving

    # --- Feature Store ---
    "feast>=0.38.0",              # Feature store — offline + online

    # --- Monitoring & Drift Detection ---
    "evidently>=0.4.0",           # PSI, CSI, data drift reports
    "scipy>=1.13.0",              # KS statistic, statistical tests

    # --- Utilities ---
    "python-dotenv>=1.0.0",       # Environment variable management
    "loguru>=0.7.0",              # Structured logging
    "tenacity>=8.3.0",            # Retry logic for external API calls
    "joblib>=1.4.0",              # Parallel feature computation + model persistence
    "tqdm>=4.66.0",               # Progress bars for batch jobs
    "typer>=0.12.0",              # CLI for training / scoring pipelines
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",              # TestClient for FastAPI
    "ruff>=0.4.0",                # Linter + formatter
    "mypy>=1.10.0",               # Static type checking
    "pre-commit>=3.7.0",          # Git hooks
    "ipykernel>=6.29.0",          # Jupyter notebooks for EDA
    "matplotlib>=3.8.0",          # Plots: ROC, reliability diagram, feature importance
    "seaborn>=0.13.0",            # Heatmaps, distribution plots
    "plotly>=5.22.0",             # Interactive charts for dashboards
]

notebooks = [
    "jupyterlab>=4.2.0",
    "ipywidgets>=8.1.0",          # Interactive widgets (SHAP plots use these)
]

[project.scripts]
credit-train  = "credit_risk.pipeline.train:main"
credit-score  = "credit_risk.pipeline.score:main"
credit-serve  = "credit_risk.api.main:serve"
credit-monitor = "credit_risk.monitoring.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/credit_risk"]

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src/credit_risk --cov-report=term-missing"
